refactor(apidom-ls): rename oas31 server-variables
