+++ packages/apidom-ls/src/config/openapi/server-variables/documentation.ts
