+++ packages/apidom-ls/src/config/openapi/server-variables/meta.ts
